Add shared domain models
