--- CatalogService/src/main/java/org/mybatis/jpetstore/domain/Item.java
@@ -1,145 +1,0 @@
-/*
- *    Copyright 2010-2021 the original author or authors.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-package org.mybatis.jpetstore.domain;
-
-import java.io.Serializable;
-import java.math.BigDecimal;
-
-/**
- * The Class Item.
- *
- * @author Eduardo Macarron
- */
-public class Item implements Serializable {
-
-  private static final long serialVersionUID = -2159121673445254631L;
-
-  private String itemId;
-  private String productId;
-  private BigDecimal listPrice;
-  private BigDecimal unitCost;
-  private int supplierId;
-  private String status;
-  private String attribute1;
-  private String attribute2;
-  private String attribute3;
-  private String attribute4;
-  private String attribute5;
-  private Product product;
-  private int quantity;
-
-  public String getItemId() {
-    return itemId;
-  }
-
-  public void setItemId(String itemId) {
-    this.itemId = itemId.trim();
-  }
-
-  public int getQuantity() {
-    return quantity;
-  }
-
-  public void setQuantity(int quantity) {
-    this.quantity = quantity;
-  }
-
-  public Product getProduct() {
-    return product;
-  }
-
-  public void setProduct(Product product) {
-    this.product = product;
-  }
-
-  public int getSupplierId() {
-    return supplierId;
-  }
-
-  public void setSupplierId(int supplierId) {
-    this.supplierId = supplierId;
-  }
-
-  public BigDecimal getListPrice() {
-    return listPrice;
-  }
-
-  public void setListPrice(BigDecimal listPrice) {
-    this.listPrice = listPrice;
-  }
-
-  public BigDecimal getUnitCost() {
-    return unitCost;
-  }
-
-  public void setUnitCost(BigDecimal unitCost) {
-    this.unitCost = unitCost;
-  }
-
-  public String getStatus() {
-    return status;
-  }
-
-  public void setStatus(String status) {
-    this.status = status;
-  }
-
-  public String getAttribute1() {
-    return attribute1;
-  }
-
-  public void setAttribute1(String attribute1) {
-    this.attribute1 = attribute1;
-  }
-
-  public String getAttribute2() {
-    return attribute2;
-  }
-
-  public void setAttribute2(String attribute2) {
-    this.attribute2 = attribute2;
-  }
-
-  public String getAttribute3() {
-    return attribute3;
-  }
-
-  public void setAttribute3(String attribute3) {
-    this.attribute3 = attribute3;
-  }
-
-  public String getAttribute4() {
-    return attribute4;
-  }
-
-  public void setAttribute4(String attribute4) {
-    this.attribute4 = attribute4;
-  }
-
-  public String getAttribute5() {
-    return attribute5;
-  }
-
-  public void setAttribute5(String attribute5) {
-    this.attribute5 = attribute5;
-  }
-
-  @Override
-  public String toString() {
-    return "(" + getItemId() + "-" + getProduct().getProductId() + ")";
-  }
-
-}
