+++ CommonLibrary/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'java-library'
+}
+
+group = 'org.mybatis'
+version = '1.0.0'
+archivesBaseName = 'jpetstore-common'
