+++ CommonLibrary/src/main/java/org/mybatis/jpetstore/domain/Product.java
@@ -14,28 +14,38 @@
  *    limitations under the License.
  */
 package org.mybatis.jpetstore.domain;
+// Shared domain model used by multiple services
 
 import java.io.Serializable;
 
 /**
- * The Class Category.
+ * The Class Product.
  *
  * @author Eduardo Macarron
  */
-public class Category implements Serializable {
+public class Product implements Serializable {
 
-  private static final long serialVersionUID = 3992469837058393712L;
+  private static final long serialVersionUID = -7492639752670189553L;
 
+  private String productId;
   private String categoryId;
   private String name;
   private String description;
+
+  public String getProductId() {
+    return productId;
+  }
+
+  public void setProductId(String productId) {
+    this.productId = productId.trim();
+  }
 
   public String getCategoryId() {
     return categoryId;
   }
 
   public void setCategoryId(String categoryId) {
-    this.categoryId = categoryId.trim();
+    this.categoryId = categoryId;
   }
 
   public String getName() {
@@ -56,7 +66,7 @@ public class Category implements Serializable {
 
   @Override
   public String toString() {
-    return getCategoryId();
+    return getName();
   }
 
 }
