+++ OrderService/build.gradle
@@ -4,5 +4,6 @@ archivesBaseName = 'jpetstore-order'
 
 dependencies {
     implementation 'org.springframework.kafka:spring-kafka:3.3.7'
+    implementation project(':CommonLibrary')
 }
