--- OrderService/src/main/java/org/mybatis/jpetstore/domain/Account.java
@@ -1,192 +1,0 @@
-/*
- *    Copyright 2010-2021 the original author or authors.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-package org.mybatis.jpetstore.domain;
-
-import java.io.Serializable;
-
-/**
- * The Class Account.
- *
- * @author Eduardo Macarron
- */
-public class Account implements Serializable {
-
-  private static final long serialVersionUID = 8751282105532159742L;
-
-  private String username;
-  private String password;
-  private String email;
-  private String firstName;
-  private String lastName;
-  private String status;
-  private String address1;
-  private String address2;
-  private String city;
-  private String state;
-  private String zip;
-  private String country;
-  private String phone;
-  private String favouriteCategoryId;
-  private String languagePreference;
-  private boolean listOption;
-  private boolean bannerOption;
-  private String bannerName;
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
-  public void setPassword(String password) {
-    this.password = password;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
-  }
-
-  public String getFirstName() {
-    return firstName;
-  }
-
-  public void setFirstName(String firstName) {
-    this.firstName = firstName;
-  }
-
-  public String getLastName() {
-    return lastName;
-  }
-
-  public void setLastName(String lastName) {
-    this.lastName = lastName;
-  }
-
-  public String getStatus() {
-    return status;
-  }
-
-  public void setStatus(String status) {
-    this.status = status;
-  }
-
-  public String getAddress1() {
-    return address1;
-  }
-
-  public void setAddress1(String address1) {
-    this.address1 = address1;
-  }
-
-  public String getAddress2() {
-    return address2;
-  }
-
-  public void setAddress2(String address2) {
-    this.address2 = address2;
-  }
-
-  public String getCity() {
-    return city;
-  }
-
-  public void setCity(String city) {
-    this.city = city;
-  }
-
-  public String getState() {
-    return state;
-  }
-
-  public void setState(String state) {
-    this.state = state;
-  }
-
-  public String getZip() {
-    return zip;
-  }
-
-  public void setZip(String zip) {
-    this.zip = zip;
-  }
-
-  public String getCountry() {
-    return country;
-  }
-
-  public void setCountry(String country) {
-    this.country = country;
-  }
-
-  public String getPhone() {
-    return phone;
-  }
-
-  public void setPhone(String phone) {
-    this.phone = phone;
-  }
-
-  public String getFavouriteCategoryId() {
-    return favouriteCategoryId;
-  }
-
-  public void setFavouriteCategoryId(String favouriteCategoryId) {
-    this.favouriteCategoryId = favouriteCategoryId;
-  }
-
-  public String getLanguagePreference() {
-    return languagePreference;
-  }
-
-  public void setLanguagePreference(String languagePreference) {
-    this.languagePreference = languagePreference;
-  }
-
-  public boolean isListOption() {
-    return listOption;
-  }
-
-  public void setListOption(boolean listOption) {
-    this.listOption = listOption;
-  }
-
-  public boolean isBannerOption() {
-    return bannerOption;
-  }
-
-  public void setBannerOption(boolean bannerOption) {
-    this.bannerOption = bannerOption;
-  }
-
-  public String getBannerName() {
-    return bannerName;
-  }
-
-  public void setBannerName(String bannerName) {
-    this.bannerName = bannerName;
-  }
-
-}
