+++ build.gradle
@@ -3,11 +3,10 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.4' apply false
 }
 
+def commonLibName = 'CommonLibrary'
+
 subprojects {
     apply plugin: 'java'
-    apply plugin: 'war'
-    apply plugin: 'org.springframework.boot'
-    apply plugin: 'io.spring.dependency-management'
 
     sourceCompatibility = JavaVersion.VERSION_21
     targetCompatibility = JavaVersion.VERSION_21
@@ -19,6 +18,12 @@ subprojects {
     ext {
         hsqldbVersion = '2.7.3'
     }
+}
+
+configure(subprojects.findAll { it.name != commonLibName }) {
+    apply plugin: 'war'
+    apply plugin: 'org.springframework.boot'
+    apply plugin: 'io.spring.dependency-management'
 
     dependencies {
         implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -37,5 +42,8 @@ subprojects {
             attributes('Dependencies': 'jdk.unsupported')
         }
     }
+}
 
+project(':CommonLibrary') {
+    apply plugin: 'java-library'
 }
